MockPrimitives (#1473)
* export FakePrimitives from fakes

* export FakePrimitives
