+++ test_vm/src/lib.rs
@@ -1,6 +1,3 @@
-use crate::fakes::FakePrimitives;
-
-use anyhow::Error;
 use cid::multihash::Code;
 use cid::Cid;
 use fil_actor_account::State as AccountState;
@@ -30,18 +27,9 @@ use fvm_ipld_hamt::{BytesKey, Hamt, Sha256};
 use fvm_shared::address::Address;
 use fvm_shared::bigint::Zero;
 use fvm_shared::clock::ChainEpoch;
-use fvm_shared::consensus::ConsensusFault;
-use fvm_shared::crypto::hash::SupportedHashes;
-use fvm_shared::crypto::signature::{
-    Signature, SECP_PUB_LEN, SECP_SIG_LEN, SECP_SIG_MESSAGE_HASH_SIZE,
-};
 use fvm_shared::econ::TokenAmount;
 use fvm_shared::error::ExitCode;
-use fvm_shared::piece::PieceInfo;
-use fvm_shared::sector::{
-    AggregateSealVerifyProofAndInfos, RegisteredSealProof, ReplicaUpdateInfo, SealVerifyInfo,
-    StoragePower, WindowPoStVerifyInfo,
-};
+use fvm_shared::sector::StoragePower;
 use fvm_shared::version::NetworkVersion;
 use fvm_shared::{MethodNum, METHOD_SEND};
 use serde::ser;
@@ -49,19 +37,18 @@ use std::cell::{RefCell, RefMut};
 use std::collections::{BTreeMap, HashMap};
 use std::rc::Rc;
 use vm_api::trace::InvocationTrace;
-use vm_api::{new_actor, ActorState, MessageResult, VMError, VM};
+use vm_api::{new_actor, ActorState, MessageResult, MockPrimitives, VMError, VM};
 
 use vm_api::util::{get_state, serialize_ok};
 
 mod constants;
 pub use constants::*;
-pub mod fakes;
 mod messaging;
 pub use messaging::*;
 
 /// An in-memory rust-execution VM for testing builtin-actors that yields sensible stack traces and debug info
 pub struct TestVM {
-    pub primitives: RefCell<FakePrimitives>,
+    pub primitives: FakePrimitives,
     pub store: Rc<MemoryBlockstore>,
     pub state_root: RefCell<Cid>,
     actors_dirty: RefCell<bool>,
@@ -85,7 +72,7 @@ impl TestVM {
             );
 
         TestVM {
-            primitives: RefCell::new(FakePrimitives::default()),
+            primitives: FakePrimitives::default(),
             store,
             state_root: RefCell::new(actors.flush().unwrap()),
             circulating_supply: RefCell::new(TokenAmount::zero()),
@@ -400,8 +387,8 @@ impl VM for TestVM {
         self.actors_dirty.replace(true);
     }
 
-    fn primitives(&self) -> Box<dyn Primitives> {
-        Box::new(self.primitives.borrow().clone())
+    fn primitives(&self) -> &dyn Primitives {
+        &self.primitives
     }
 
     fn actor_manifest(&self) -> BTreeMap<Cid, Type> {
@@ -451,62 +438,7 @@ impl VM for TestVM {
         self.timestamp.replace(timestamp);
     }
 
-    fn override_hash_blake2b(&self, f: fn(&[u8]) -> [u8; 32]) {
-        self.primitives.borrow_mut().hash_blake2b = Some(f);
-    }
-
-    fn override_hash(&self, f: fn(SupportedHashes, &[u8]) -> Vec<u8>) {
-        self.primitives.borrow_mut().hash = Some(f);
-    }
-
-    fn override_hash_64(&self, f: fn(SupportedHashes, &[u8]) -> ([u8; 64], usize)) {
-        self.primitives.borrow_mut().hash_64 = Some(f);
-    }
-
-    fn override_compute_unsealed_sector_cid(
-        &self,
-        f: fn(RegisteredSealProof, &[PieceInfo]) -> Result<Cid, Error>,
-    ) {
-        self.primitives.borrow_mut().compute_unsealed_sector_cid = Some(f);
-    }
-
-    fn override_recover_secp_public_key(
-        &self,
-        f: fn(
-            &[u8; SECP_SIG_MESSAGE_HASH_SIZE],
-            &[u8; SECP_SIG_LEN],
-        ) -> Result<[u8; SECP_PUB_LEN], Error>,
-    ) {
-        self.primitives.borrow_mut().recover_secp_public_key = Some(f);
-    }
-
-    fn override_verify_post(&self, f: fn(&WindowPoStVerifyInfo) -> Result<(), Error>) {
-        self.primitives.borrow_mut().verify_post = Some(f);
-    }
-
-    fn override_verify_consensus_fault(
-        &self,
-        f: fn(&[u8], &[u8], &[u8]) -> Result<Option<ConsensusFault>, Error>,
-    ) {
-        self.primitives.borrow_mut().verify_consensus_fault = Some(f);
-    }
-
-    fn override_batch_verify_seals(&self, f: fn(&[SealVerifyInfo]) -> Result<Vec<bool>, Error>) {
-        self.primitives.borrow_mut().batch_verify_seals = Some(f);
-    }
-
-    fn override_verify_aggregate_seals(
-        &self,
-        f: fn(&AggregateSealVerifyProofAndInfos) -> Result<(), Error>,
-    ) {
-        self.primitives.borrow_mut().verify_aggregate_seals = Some(f);
-    }
-
-    fn override_verify_signature(&self, f: fn(&Signature, &Address, &[u8]) -> Result<(), Error>) {
-        self.primitives.borrow_mut().verify_signature = Some(f);
-    }
-
-    fn override_verify_replica_update(&self, f: fn(&ReplicaUpdateInfo) -> Result<(), Error>) {
-        self.primitives.borrow_mut().verify_replica_update = Some(f);
+    fn mut_primitives(&self) -> &dyn MockPrimitives {
+        &self.primitives
     }
 }
